requirements.txt
browser-use
python-dotenv
openai
pandas
reportlab
fpdf2
matplotlib
Pillow
